Add a slack role and dependent snap role
This commit adds slack as a default tool on my dev machine. Since we're
installing slack via the snap tool, it also adds a dependent snap role
to ensure that snap is configured on the machine.

Unfortunately slack still requires classic isolation which gives it just
as much power as a locally installed deb, but at least this way we don't
need to manage the ppa.