Add SKIP_ANSIBLE_INSTALL environment var
As of Ubuntu 20.04 in late 2020 there is a bug due to a patched version
of python virtualenv which is preventing the ansible project from
supporting Ubuntu 20.04 as a valid PPA. As a result the bootstrap
process fails. This means I need to manually disable the check until the
PPA is fixed.

This commit allows me to do so by setting the environment variable
SKIP_ANSIBLE_INSTALL to any value prior to runnign run.sh