dotnet - use ubuntu official packages
Microsoft dotnet and its SDK/runtime/targeting packs are now officially
available in the ubuntu package repository removing the need for the
Microsoft specific package repository. This comes with some caviats.

The packages conflict so during migration cleanup or re-install may be
necessary after purging the MS package repository or packages from your
system via:

    apt remove --purge 'dotnet*'
    apt remove --purge 'aspnetcore*'
    apt remove --purge 'netstandard*'
    apt remove --purge packages-microsoft-prod

Following the above and a package autoremove it should be safe to
re-install dotnet per this ansible script.

New installs need not be concerned.