Swap back to legacy apt repo oneline format for 1password
The 1password package updates its own package repository reference as
part of its package install. This means that if we use deb822 format
that we will have duplicate sources and different key formats. This
prevents apt from running.

This commit swaps back to the form used by 1password in order to avoid
conflicts.